+++ packages/frontend/src/features/omnibar/api/search.ts
@@ -1,15 +1,24 @@
-import { SearchResults } from '@lightdash/common';
+import { SearchFilters, SearchResults } from '@lightdash/common';
 import { lightdashApi } from '../../../api';
 
 export const getSearchResults = async ({
     projectUuid,
     query,
+    filters,
 }: {
     projectUuid: string;
     query: string;
-}) =>
-    lightdashApi<SearchResults>({
-        url: `/projects/${projectUuid}/search/${encodeURIComponent(query)}`,
+    filters?: SearchFilters;
+}) => {
+    const searchParams = filters
+        ? new URLSearchParams(filters).toString()
+        : undefined;
+
+    return lightdashApi<SearchResults>({
+        url: `/projects/${projectUuid}/search/${encodeURIComponent(query)}${
+            searchParams ? `?${searchParams}` : ''
+        }`,
         method: 'GET',
         body: undefined,
     });
+};