+++ packages/frontend/src/features/omnibar/hooks/useSearch.ts
@@ -1,20 +1,22 @@
-import { ApiError, SearchResults } from '@lightdash/common';
+import { ApiError, SearchFilters, SearchResults } from '@lightdash/common';
 import { useQuery } from '@tanstack/react-query';
 import { getSearchResults } from '../api/search';
+import { SearchResultMap } from '../types/searchResultMap';
+import { getSearchItemMap } from '../utils/getSearchItemMap';
 
 export const OMNIBAR_MIN_QUERY_LENGTH = 3;
 
-const useSearch = (projectUuid: string, query: string = '') => {
-    return useQuery<SearchResults, ApiError>({
-        queryKey: ['search', query],
-        queryFn: () =>
-            getSearchResults({
-                projectUuid,
-                query,
-            }),
+const useSearch = (
+    projectUuid: string,
+    query: string = '',
+    filters?: SearchFilters,
+) =>
+    useQuery<SearchResults, ApiError, SearchResultMap>({
+        queryKey: ['search', filters?.type ?? 'all', query],
+        queryFn: () => getSearchResults({ projectUuid, query, filters }),
         retry: false,
         enabled: query.length >= OMNIBAR_MIN_QUERY_LENGTH,
+        select: (data) => getSearchItemMap(data, projectUuid),
     });
-};
 
 export default useSearch;